# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "API Description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
        "/api/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get post comments",
                "description": "Returns a paginated list of post comments. Results can be filtered by comment status.",
                "operationId": "aa239f69b4c682985441ea2c62960ae1",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter comments by status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "draft",
                                "published",
                                "taken_down"
                            ],
                            "example": "published"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagination page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Number of comments per page. Maximum value is 100.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "example": 25
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated comments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "commentedBy": {
                                                        "type": "string",
                                                        "example": "Dr. Magdalen Walsh"
                                                    },
                                                    "comment": {
                                                        "type": "string",
                                                        "example": "This is a post comment."
                                                    },
                                                    "post": {
                                                        "type": "integer",
                                                        "example": 27
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "published"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 1000
                                                },
                                                "page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "pageSize": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "lastPage": {
                                                    "type": "integer",
                                                    "example": 40
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/stats/overview": {
            "get": {
                "tags": [
                    "Stats"
                ],
                "summary": "Get overview statistics",
                "operationId": "a084da48d55b0fce59cbf3db16911e31",
                "responses": {
                    "200": {
                        "description": "Overview statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "totalUsers": {
                                            "type": "integer",
                                            "example": 1200
                                        },
                                        "activeUsers": {
                                            "type": "integer",
                                            "example": 950
                                        },
                                        "pendingReports": {
                                            "type": "integer",
                                            "example": 35
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/stats/geographic-distribution": {
            "get": {
                "tags": [
                    "Stats"
                ],
                "summary": "Get user geographical distribution grouped by country",
                "operationId": "3bfd2fa8e4b101bc86d2cff187518d10",
                "responses": {
                    "200": {
                        "description": "Geographical distribution grouped by country.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "country": {
                                                "type": "string",
                                                "example": "Italy"
                                            },
                                            "count": {
                                                "type": "integer",
                                                "example": 223
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/stats/viral-posts": {
            "get": {
                "tags": [
                    "Stats"
                ],
                "summary": "Get user growth grouped by period",
                "operationId": "bc8cad323957ceee9c86b1810f3f5e33",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Start date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-05-01"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "End date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-05-31"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Viral posts for a given date range.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "viralScore": {
                                                "type": "number",
                                                "format": "double",
                                                "example": 1251.8
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/stats/user-growth": {
            "get": {
                "tags": [
                    "Stats"
                ],
                "summary": "Get user growth grouped by period",
                "operationId": "aa4278438a580776c71081a1799132e6",
                "parameters": [
                    {
                        "name": "groupBy",
                        "in": "query",
                        "description": "Grouping period",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "day",
                                "week",
                                "month",
                                "year"
                            ],
                            "example": "month"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Start date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-05-01"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "End date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-05-31"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User growth data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "period": {
                                                "type": "string",
                                                "example": "2026-05"
                                            },
                                            "count": {
                                                "type": "integer",
                                                "example": 1200
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Comments",
            "description": "Comments"
        },
        {
            "name": "Stats",
            "description": "Stats"
        }
    ]
}